test(lance): move composite transaction tests into the library
A dedicated integration test gets its own binary, which links every
dependency of the crate. Fold the six tests into the existing
`dataset::tests::dataset_transactions` module as a `composite` submodule and
delete the standalone target.
